修复添加菜品设置菜品价格只能为整数的bug

+++ code/pages/addFood/addFood.js
@@ -83,7 +83,7 @@ Page({
             countNum: 0
         }
         model.MenuItemName = this.data.name
-        model.Price =  parseInt(this.data.price)
+        model.Price =  parseFloat(this.data.price)
         model.ItemDescription = this.data.describe
 
         this.setData({
